TEIID-4441: make ftps work, merge vsffile and file translator to one
