+++ connectors/file/connector-ftp/src/main/resources/org/teiid/resource/adapter/ftp/i18n.properties
@@ -33,6 +33,8 @@ ftp_client_port=Port number should be > 0
 ftp_client_invalid_array=Invalid {0}, the {1} should be comma separate string list
 ftp_client_clientMode=Only local modes are supported. Was: {0}
 ftp_connect_failed=Connecting to server {0}:{1} failed. Please check the connection.
+ftp_certificate_path=Invalidate certificate path {0} {1}
+ftp_ketstore_path=Invalidate keystore path {0} {1}
 ftp_login_failed=Login failed. The response from the server is: {0} 
 ftp_failed_read=Failed to read from {0}. Server replied with: {1}
 ftp_failed_write=Failed to write to {0}. Server replied with: {1}
