+++ connectors/file/translator-file/src/main/resources/META-INF/services/org.teiid.translator.ExecutionFactory
@@ -1,2 +1,1 @@
 org.teiid.translator.file.FileExecutionFactory
-org.teiid.translator.file.VirtualExecutionFactory
