+++ teiid-feature-pack/wildfly-integration-feature-pack/src/main/resources/content/bin/scripts/teiid-domain-mode-install.cli
@@ -109,7 +109,6 @@ connect
 /profile=ha/subsystem=teiid/translator=intersystems-cache:add(module=org.jboss.teiid.translator.jdbc)
 /profile=ha/subsystem=teiid/translator=netezza:add(module=org.jboss.teiid.translator.jdbc)
 /profile=ha/subsystem=teiid/translator=file:add(module=org.jboss.teiid.translator.file)
-/profile=ha/subsystem=teiid/translator=vfsfile:add(module=org.jboss.teiid.translator.file)
 /profile=ha/subsystem=teiid/translator=ldap:add(module=org.jboss.teiid.translator.ldap)
 /profile=ha/subsystem=teiid/translator=loopback:add(module=org.jboss.teiid.translator.loopback)
 /profile=ha/subsystem=teiid/translator=olap:add(module=org.jboss.teiid.translator.olap)
@@ -190,7 +189,6 @@ connect
 /profile=full-ha/subsystem=teiid/translator=intersystems-cache:add(module=org.jboss.teiid.translator.jdbc)
 /profile=full-ha/subsystem=teiid/translator=netezza:add(module=org.jboss.teiid.translator.jdbc)
 /profile=full-ha/subsystem=teiid/translator=file:add(module=org.jboss.teiid.translator.file)
-/profile=full-ha/subsystem=teiid/translator=vfsfile:add(module=org.jboss.teiid.translator.file)
 /profile=full-ha/subsystem=teiid/translator=ldap:add(module=org.jboss.teiid.translator.ldap)
 /profile=full-ha/subsystem=teiid/translator=loopback:add(module=org.jboss.teiid.translator.loopback)
 /profile=full-ha/subsystem=teiid/translator=olap:add(module=org.jboss.teiid.translator.olap)
